rpi-motor-sensor-control/requirements.txt
RPi.GPIO
gpiozero
numpy
opencv-python-headless